the backend for ifinance app

+++ routers/stockRouter.js
@@ -1,35 +1,44 @@
 const router = require("express").Router();
 const StockData = require("../models/stocksModel");
+const auth = require("../middleware/auth");
+
+const https = require("https");
+const url = require("url");
 
 
-// ================ Get Stocks from MyPortfolio =============== //
+// ================ Get Stocks from MyPortfolio with user Auth =============== //
 
-router.get("/", async (req, res) => {
+router.get("/", auth, async (req, res) => {
     try {
-        const portfolio = await StockData.find();
-        console.log(portfolio);
-        res.send(portfolio);
+        console.log(req.user);
+        const portfolio = await StockData.find({user: req.user });
+        res.json(portfolio);
     }
     catch(err) {
         res.status(500).send();
     }
 });
 
-// ================ Add Stock to MyPortfolio =============== //
+// ================ Add Stock to MyPortfolio with user Auth =============== //
 
-router.post("/", async (req, res) => {
+router.post("/", auth, async (req, res) => {
     try {
-        const {ticker, company, price, description} = req.body;
+        const {symbol, company, priceAtPurchase, sharesPurchased, description} = req.body;
         console.log(req.body);
 
         //VALIDATION
 
-        if (!ticker && !company && !price) {
+        if (!symbol && !company && !priceAtPurchase) {
             return res.status(400).json({ errorMessage: "Please enter a stock Ticker, Company Name, and Price."});
         }
 
         const addStockInfo = new StockData({
-            ticker, company, price, description
+            symbol,
+            company,
+            priceAtPurchase,
+            sharesPurchased,
+            description,
+            user: req.user,
         });
 
         // once PROMISE has been resolved, pause..., then continue
@@ -44,30 +53,35 @@ router.post("/", async (req, res) => {
 
 // ================ UPDATE Stocks from MyPortfolio =============== //
 
-router.put("/:id", async (req, res) => {
+router.put("/:id", auth, async (req, res) => {
     try {
-        const {ticker, company, price, description} = req.body;
+        const {symbol, company, priceAtPurchase, sharesPurchased, description} = req.body;
         const stockId = req.params.id;
         console.log(stockId);
 
         // VALIDATION ... what if there isnt an ID selected. catches potential front end error. //
 
-        if (!ticker && !company && !price) {
+        if (!symbol && !company) {
             return res.status(400).json({ errorMessage: "Please enter a stock Ticker, Company Name, and Price."});
         }
 
-        if(!stockId) {
+        if (!stockId) {
             return res.status(400).json({ errorMessage: "Stock ID not found...please contact support"});
         }
         const originalStock = await StockData.findById(stockId);
 
-        if(!originalStock) {
+        if (!originalStock) {
             return res.status(400).json({ errorMessage: "Stock ID not found... please contact developer"});
         }
 
-       originalStock.ticker = ticker;
+        if (originalStock.user.toString() !== req.user) {
+            return res.status(401).json({ errorMessage: "Unauthorized."});
+        }
+
+       originalStock.symbol = symbol;
        originalStock.company = company;
-       originalStock.price = price;
+       originalStock.priceAtPurchase = priceAtPurchase;
+       originalStock.sharesPurchased = sharesPurchased;
        originalStock.description = description;
 
        const updatedStock = await originalStock.save();
@@ -83,12 +97,12 @@ router.put("/:id", async (req, res) => {
 
 // ================ DELETE Stocks from MyPortfolio =============== //
 
-router.delete("/:id", async (req, res) => {
+router.delete("/:id", auth, async (req, res) => {
     try {
         const stockId = req.params.id;
         console.log(stockId);
 
-        // VALIDATION ... what if there isnt an ID selected. catches potential front end error. //
+    // VALIDATION ... what if there isnt an ID selected. catches potential front end error. //
 
         if(!stockId) {
             return res.status(400).json({ errorMessage: "Stock ID not found...please contact support"});
@@ -98,10 +112,16 @@ router.delete("/:id", async (req, res) => {
         if(!existingStock) {
             return res.status(400).json({ errorMessage: "Stock ID not found... please contact developer"});
         }
-        // delete stock
+
+    // check if user is logged in before delete
+        if (existingStock.user.toString() !== req.user) {
+            return res.status(401).json({ errorMessage: "Unauthorized."});
+        }
+
+    // delete stock
         await existingStock.delete();
 
-        //tell user which stock has been deleted
+    //tell user which stock has been deleted
         res.json(existingStock);
 
     }
